+++ tests/unit/test_app.py
@@ -53,46 +53,6 @@ def create_request_with_content_type(content_type):
     )
 
 
-def create_event(uri, method, path, content_type='application/json'):
-    return {
-        'requestContext': {
-            'httpMethod': method,
-            'resourcePath': uri,
-        },
-        'headers': {
-            'Content-Type': content_type,
-        },
-        'pathParameters': path,
-        'queryStringParameters': {},
-        'body': None,
-        'stageVariables': {},
-    }
-
-
-def create_empty_header_event(uri, method, path,
-                              content_type='application/json'):
-    return {
-        'requestContext': {
-            'httpMethod': method,
-            'resourcePath': uri,
-        },
-        'headers': None,
-        'pathParameters': path,
-        'queryStringParameters': {},
-        'body': "",
-        'stageVariables': {},
-    }
-
-
-def create_event_with_body(body, uri='/', method='POST',
-                           content_type='application/json'):
-    event = create_event(uri, method, {}, content_type)
-    if content_type == 'application/json':
-        body = json.dumps(body)
-    event['body'] = body
-    return event
-
-
 def assert_response_body_is(response, body):
     assert json.loads(response['body']) == body
 
@@ -221,13 +181,13 @@ def test_error_on_unknown_event(sample_app):
     assert json_response_body(raw_response)['Code'] == 'InternalServerError'
 
 
-def test_can_route_api_call_to_view_function(sample_app):
+def test_can_route_api_call_to_view_function(sample_app, create_event):
     event = create_event('/index', 'GET', {})
     response = sample_app(event, context=None)
     assert_response_body_is(response, {'hello': 'world'})
 
 
-def test_can_call_to_dict_on_current_request(sample_app):
+def test_can_call_to_dict_on_current_request(sample_app, create_event):
     @sample_app.route('/todict')
     def todict():
         return sample_app.current_request.to_dict()
@@ -243,7 +203,8 @@ def test_can_call_to_dict_on_current_request(sample_app):
     assert isinstance(json.loads(json.dumps(response)), dict)
 
 
-def test_request_to_dict_does_not_contain_internal_attrs(sample_app):
+def test_request_to_dict_does_not_contain_internal_attrs(sample_app,
+                                                         create_event):
     @sample_app.route('/todict')
     def todict():
         return sample_app.current_request.to_dict()
@@ -253,34 +214,35 @@ def test_request_to_dict_does_not_contain_internal_attrs(sample_app):
     assert not internal_attrs
 
 
-def test_will_pass_captured_params_to_view(sample_app):
+def test_will_pass_captured_params_to_view(sample_app, create_event):
     event = create_event('/name/{name}', 'GET', {'name': 'james'})
     response = sample_app(event, context=None)
     response = json_response_body(response)
     assert response == {'provided-name': 'james'}
 
 
-def test_error_on_unsupported_method(sample_app):
+def test_error_on_unsupported_method(sample_app, create_event):
     event = create_event('/name/{name}', 'POST', {'name': 'james'})
     raw_response = sample_app(event, context=None)
     assert raw_response['statusCode'] == 405
     assert json_response_body(raw_response)['Code'] == 'MethodNotAllowedError'
 
 
-def test_error_on_unsupported_method_gives_feedback_on_method(sample_app):
+def test_error_on_unsupported_method_gives_feedback_on_method(sample_app,
+                                                              create_event):
     method = 'POST'
     event = create_event('/name/{name}', method, {'name': 'james'})
     raw_response = sample_app(event, context=None)
     assert 'POST' in json_response_body(raw_response)['Message']
 
 
-def test_no_view_function_found(sample_app):
+def test_no_view_function_found(sample_app, create_event):
     bad_path = create_event('/noexist', 'GET', {})
     with pytest.raises(app.ChaliceError):
         sample_app(bad_path, context=None)
 
 
-def test_can_access_context():
+def test_can_access_context(create_event):
     demo = app.Chalice('app-name')
 
     @demo.route('/index')
@@ -296,7 +258,7 @@ def test_can_access_context():
     assert result == serialized_lambda_context
 
 
-def test_can_access_raw_body():
+def test_can_access_raw_body(create_event):
     demo = app.Chalice('app-name')
 
     @demo.route('/index')
@@ -310,7 +272,7 @@ def test_can_access_raw_body():
     assert result == {'rawbody': '{"hello": "world"}'}
 
 
-def test_raw_body_cache_returns_same_result():
+def test_raw_body_cache_returns_same_result(create_event):
     demo = app.Chalice('app-name')
 
     @demo.route('/index')
@@ -330,7 +292,7 @@ def test_raw_body_cache_returns_same_result():
     assert result['rawbody'] == result['rawbody2']
 
 
-def test_can_have_views_of_same_route_but_different_methods():
+def test_can_have_views_of_same_route_but_different_methods(create_event):
     demo = app.Chalice('app-name')
 
     @demo.route('/index', methods=['GET'])
@@ -366,7 +328,7 @@ def test_error_on_duplicate_route_methods():
             return {'foo': 'bar'}
 
 
-def test_json_body_available_with_right_content_type():
+def test_json_body_available_with_right_content_type(create_event):
     demo = app.Chalice('demo-app')
 
     @demo.route('/', methods=['POST'])
@@ -381,7 +343,7 @@ def test_json_body_available_with_right_content_type():
     assert result == {'foo': 'bar'}
 
 
-def test_cant_access_json_body_with_wrong_content_type():
+def test_cant_access_json_body_with_wrong_content_type(create_event):
     demo = app.Chalice('demo-app')
 
     @demo.route('/', methods=['POST'], content_types=['application/xml'])
@@ -398,7 +360,7 @@ def test_cant_access_json_body_with_wrong_content_type():
     assert raw_body == '<Message>hello</Message>'
 
 
-def test_json_body_available_on_multiple_content_types():
+def test_json_body_available_on_multiple_content_types(create_event_with_body):
     demo = app.Chalice('demo-app')
 
     @demo.route('/', methods=['POST'],
@@ -425,7 +387,8 @@ def test_json_body_available_on_multiple_content_types():
     assert raw_body == '{"foo": "bar"}'
 
 
-def test_json_body_available_with_lowercase_content_type_key():
+def test_json_body_available_with_lowercase_content_type_key(
+        create_event_with_body):
     demo = app.Chalice('demo-app')
 
     @demo.route('/', methods=['POST'])
@@ -451,7 +414,7 @@ def test_content_types_must_be_lists():
             return {'foo': 'bar'}
 
 
-def test_content_type_validation_raises_error_on_unknown_types():
+def test_content_type_validation_raises_error_on_unknown_types(create_event):
     demo = app.Chalice('demo-app')
 
     @demo.route('/', methods=['POST'], content_types=['application/xml'])
@@ -467,7 +430,7 @@ def test_content_type_validation_raises_error_on_unknown_types():
     assert 'application/bad-xml' in json_response['Message']
 
 
-def test_content_type_with_charset():
+def test_content_type_with_charset(create_event):
     demo = app.Chalice('demo-app')
 
     @demo.route('/', content_types=['application/json'])
@@ -479,7 +442,7 @@ def test_content_type_with_charset():
     assert response == {'foo': 'bar'}
 
 
-def test_can_return_response_object():
+def test_can_return_response_object(create_event):
     demo = app.Chalice('app-name')
 
     @demo.route('/index')
@@ -493,7 +456,7 @@ def test_can_return_response_object():
                         'headers': {'Content-Type': 'application/json'}}
 
 
-def test_headers_have_basic_validation():
+def test_headers_have_basic_validation(create_event):
     demo = app.Chalice('app-name')
 
     @demo.route('/index')
@@ -509,7 +472,7 @@ def test_headers_have_basic_validation():
     assert json.loads(response['body'])['Code'] == 'InternalServerError'
 
 
-def test_empty_headers_have_basic_validation():
+def test_empty_headers_have_basic_validation(create_empty_header_event):
     demo = app.Chalice('app-name')
 
     @demo.route('/index')
@@ -522,7 +485,7 @@ def test_empty_headers_have_basic_validation():
     assert response['statusCode'] == 200
 
 
-def test_no_content_type_is_still_allowed():
+def test_no_content_type_is_still_allowed(create_event):
     # When the content type validation happens in API gateway, it appears
     # to assume a default of application/json, so the chalice handler needs
     # to emulate that behavior.
@@ -540,7 +503,7 @@ def test_no_content_type_is_still_allowed():
     assert json_response == {'success': True}
 
 
-def test_can_base64_encode_binary_media_types_bytes():
+def test_can_base64_encode_binary_media_types_bytes(create_event):
     demo = app.Chalice('demo-app')
 
     @demo.route('/index')
@@ -559,7 +522,8 @@ def test_can_base64_encode_binary_media_types_bytes():
     assert response['headers']['Content-Type'] == 'application/octet-stream'
 
 
-def test_can_return_text_even_with_binary_content_type_configured():
+def test_can_return_text_even_with_binary_content_type_configured(
+        create_event):
     demo = app.Chalice('demo-app')
 
     @demo.route('/index')
@@ -614,7 +578,7 @@ def test_route_inequality(view_function):
     assert not a == b
 
 
-def test_exceptions_raised_as_chalice_errors(sample_app):
+def test_exceptions_raised_as_chalice_errors(sample_app, create_event):
 
     @sample_app.route('/error')
     def raise_error():
@@ -630,7 +594,7 @@ def test_exceptions_raised_as_chalice_errors(sample_app):
     assert raw_response['statusCode'] == 500
 
 
-def test_original_exception_raised_in_debug_mode(sample_app):
+def test_original_exception_raised_in_debug_mode(sample_app, create_event):
     sample_app.debug = True
 
     @sample_app.route('/error')
@@ -646,7 +610,8 @@ def test_original_exception_raised_in_debug_mode(sample_app):
     assert 'You will see this error' in response['body']
 
 
-def test_chalice_view_errors_propagate_in_non_debug_mode(sample_app):
+def test_chalice_view_errors_propagate_in_non_debug_mode(sample_app,
+                                                         create_event):
     @sample_app.route('/notfound')
     def notfound():
         raise NotFoundError("resource not found")
@@ -657,7 +622,7 @@ def test_chalice_view_errors_propagate_in_non_debug_mode(sample_app):
     assert json_response_body(raw_response)['Code'] == 'NotFoundError'
 
 
-def test_chalice_view_errors_propagate_in_debug_mode(sample_app):
+def test_chalice_view_errors_propagate_in_debug_mode(sample_app, create_event):
     @sample_app.route('/notfound')
     def notfound():
         raise NotFoundError("resource not found")
@@ -678,7 +643,7 @@ def test_case_insensitive_mapping():
     assert repr({'header': 'Value'}) in repr(mapping)
 
 
-def test_unknown_kwargs_raise_error(sample_app):
+def test_unknown_kwargs_raise_error(sample_app, create_event):
     with pytest.raises(TypeError):
         @sample_app.route('/foo', unknown_kwargs='foo')
         def badkwargs():
@@ -731,7 +696,7 @@ def test_json_body_available_when_content_type_matches(content_type, is_json):
         assert request.json_body is None
 
 
-def test_can_receive_binary_data():
+def test_can_receive_binary_data(create_event_with_body):
     content_type = 'application/octet-stream'
     demo = app.Chalice('demo-app')
 
@@ -753,7 +718,8 @@ def test_can_receive_binary_data():
     assert response['body'] == body
 
 
-def test_cannot_receive_base64_string_with_binary_response():
+def test_cannot_receive_base64_string_with_binary_response(
+        create_event_with_body):
     content_type = 'application/octet-stream'
     demo = app.Chalice('demo-app')
 
@@ -1248,7 +1214,7 @@ def test_aws_execution_env_set():
     )
 
 
-def test_can_use_out_of_order_args():
+def test_can_use_out_of_order_args(create_event):
     demo = app.Chalice('demo-app')
 
     # Note how the url params and function args are out of order.